+++ packages/cozy-scanner/src/DocumentTypeDataHelpers.spec.jsx
@@ -1,21 +1,13 @@
-import { getItemsByCategory, getThemeByItem } from './DocumentTypeDataHelpers'
+import { getThemeByItem } from './DocumentTypeDataHelpers'
 
 describe('DocumentTypeDataHelpers', () => {
-  it('should test getItemByCategory', () => {
-    const items = getItemsByCategory({ label: 'identity' })
-    expect(items.length).toBe(6)
-  })
-  it('should return null if no cate with this label', () => {
-    const items = getItemsByCategory('')
-    expect(items).toBeNull()
-  })
-
   it('should test getThemeByItem and return defaultTheme', () => {
     const item = {
-      id: '4',
-      classification: 'identity_document',
-      subClassification: 'family_record_book',
       label: 'family_record_book',
+      purpose: 'attestation',
+      sourceCategory: 'gov',
+      sourceSubCategory: 'civil_registration',
+      subjects: ['family'],
       defaultTheme: 'theme2'
     }
     const theme = getThemeByItem(item)
@@ -24,10 +16,11 @@ describe('DocumentTypeDataHelpers', () => {
 
   it('should test getThemeByItem and return the theme if no defaultTheme', () => {
     const item = {
-      id: '3',
-      classification: 'identity_document',
-      subClassification: 'residence_permit',
-      label: 'residence_permit'
+      label: 'residence_permit',
+      purpose: 'attestation',
+      sourceCategory: 'gov',
+      sourceSubCategory: 'immigration',
+      subjects: ['permit', 'identity']
     }
     const theme = getThemeByItem(item)
     expect(theme.id).toBe('theme1')